framework/requirements.txt
anyio==4.1.0
asgiref==3.7.2
attrs==23.1.0
azure-common==1.1.25
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
boto3==1.17.85
botocore==1.20.85
cachetools==4.1.0
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.4
click==8.1.3
clickclick==20.10.2
connexion==3.0.5
content-size-limit-asgi==0.1.5
cryptography==42.0.4
Cython==0.29.36
defusedxml==0.6.0
docker==6.0.0
docker-pycreds==0.4.0
docutils==0.15.2
ecdsa==0.16.1
exceptiongroup==1.2.0
frozenlist==1.2.0
future==0.18.3
google-api-core==1.30.0
googleapis-common-protos==1.51.0
google-auth==1.28.0
google-cloud-core==1.7.1
google-cloud-pubsub==2.7.1
google-cloud-storage==1.39.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
greenlet==2.0.2
grpc-google-iam-v1==0.12.3
grpcio==1.58.0
httpcore==1.0.2
httpx==0.26.0
h11==0.14.0
idna==3.7.0
importlib-metadata==6.8.0
inflection==0.3.1
Jinja2==3.1.3
jmespath==0.9.5
jsonschema==4.20.0
jsonschema-path==0.3.2
jsonschema-specifications==2023.11.2
lazy-object-proxy==1.10.0
libcst==0.3.20
MarkupSafe==2.1.2
more-itertools==8.2.0
multidict==5.2.0
mypy-extensions==0.4.3
numpy==1.26.0
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
packaging==20.9
pathable==0.4.3
pathlib==1.0.1
protobuf==3.19.6
proto-plus==1.19.0
pyarrow==14.0.1
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.8.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-json-logger==2.0.2
python-multipart==0.0.6
pytz==2020.1
PyYAML==5.4.1
referencing==0.31.1
requests==2.31.0
rfc3339-validator==0.1.4
rpds-py==0.15.2
rsa==4.7.2
s3transfer==0.4.2
secure==0.3.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.32.0.post1
tabulate==0.8.9
typing-extensions==4.5.0
typing-inspect==0.7.1
urllib3==1.26.18
uvloop==0.17.0
uvicorn==0.24.0.post1
websocket-client==0.57.0
Werkzeug==3.0.1
xmltodict==0.12.0
yarl==1.7.0
zipp==3.3.2
content_size_limit_asgi==0.1.5
